Move language state logic from WithProviders to I18nProvider
