+++ app/WithProviders.tsx
@@ -3,6 +3,7 @@ import { useState, useEffect } from "react";
 import { PageContextProvider } from "@contexts/pageContext";
 import { FlightContextProvider } from "@contexts/flightContext";
 import { DialogueContextProvider } from "@contexts/dialogueContext";
+import { UserSettingsContextProvider } from "@contexts/UserSettingsContext";
 import { I18nProvider, LOCALES } from "./i18n";
 
 type WithProvidersProps = {
@@ -10,24 +11,14 @@ type WithProvidersProps = {
 };
 
 export const WithProviders = ({ children }: WithProvidersProps) => {
-  const [lang, setLang] = useState<string>(LOCALES.JAPANESE);
-
-  useEffect(() => {
-    const locale = navigator.language;
-    const baseLocale = locale.split("-")[0];
-    if (baseLocale === "en") {
-      setLang(LOCALES.ENGLISH);
-    } else if (baseLocale === "ja") {
-      setLang(LOCALES.JAPANESE);
-    }
-  }, []);
-
   return (
     <PageContextProvider>
       <FlightContextProvider>
-        <I18nProvider locale={lang}>
-          <DialogueContextProvider>{children}</DialogueContextProvider>
-        </I18nProvider>
+        <UserSettingsContextProvider>
+          <I18nProvider>
+            <DialogueContextProvider>{children}</DialogueContextProvider>
+          </I18nProvider>
+        </UserSettingsContextProvider>
       </FlightContextProvider>
     </PageContextProvider>
   );
