+++ app/i18n/Provider.tsx
@@ -1,4 +1,5 @@
 "use client";
+import { useEffect, useState } from "react";
 import { Fragment } from "react";
 import { IntlProvider } from "react-intl";
 
@@ -7,15 +8,26 @@ import messages from "./messages";
 
 type Props = {
   children: any;
-  locale: string;
 };
 
-const Provider = ({ children, locale = LOCALES.ENGLISH }: Props) => {
+const Provider = ({ children }: Props) => {
+  const [lang, setLang] = useState<string>(LOCALES.JAPANESE);
+
+  useEffect(() => {
+    const locale = navigator.language;
+    const baseLocale = locale.split("-")[0];
+    if (baseLocale === "en") {
+      setLang(LOCALES.ENGLISH);
+    } else if (baseLocale === "ja") {
+      setLang(LOCALES.JAPANESE);
+    }
+  }, []);
+
   return (
     <IntlProvider
       textComponent={Fragment as any}
-      locale={locale}
-      messages={messages[locale]}
+      locale={lang}
+      messages={messages[lang]}
     >
       {children}
     </IntlProvider>
